backend/requirements.txt
torch
torchvision
ultralytics
opencv-python
fastapi
python-multipart
uvicorn
Pillow
